refactor: Update get_facility_free_space API endpoint
The `get_facility_free_space` API endpoint in `main_advanced.py` has been updated to include a new parameter `sam` for specifying the version of the SAM API to use. This allows for flexibility in choosing between different versions of the advanced search functionality. The `search` function in `main_advanced.py` has also been modified to handle the different versions of the SAM API.